generalize iconbutton hover style
