+++ src/frontend/src/views/Data/components/FilterPanelToggle/style.ts
@@ -1,23 +1,6 @@
 import styled from "@emotion/styled";
-import { Button } from "@material-ui/core";
 import { getColors, getSpacings } from "czifui";
 import IconFilters from "src/common/icons/IconFilters.svg";
-
-export const StyledButton = styled(Button)`
-  border-radius: 50%;
-  flex: 0 0 0;
-  min-width: 0;
-
-  ${(props) => {
-    const colors = getColors(props);
-    const spacings = getSpacings(props);
-    return `
-      fill: ${colors?.gray[500]};
-      /* margin: ${spacings?.xxxs}px;
-      padding: ${spacings?.xs}px; */
-    `;
-  }}
-`;
 
 export const StyledDiv = styled.div`
   display: flex;
